added some map checks

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: miduarte <miduarte@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/10 11:40:42 by miduarte          #+#    #+#             */
-/*   Updated: 2025/04/21 15:13:43 by miduarte         ###   ########.fr       */
+/*   Updated: 2025/07/14 12:48:01 by miduarte         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char	*str)
+int	ft_strlen(const char	*str)
 {
 	int	i;
